Add initialization check for signing
If the initialization is not complete when trying to sign in, the response from the identity-provider will be "lost" because no `Notifier` is registered with the @openid/appauth code.

This will lead to an infinite wait for the result of the login. That is quite time-consuming to debug. Therefore, an error should make things a lot easier for the programmer.