Make sure that AuthService.init() is always executed
Similar topic as already discussed: If AuthService.init() is not executed before a login-flow is started, the login-flow never completes because no listener is registered with the @openid/appauth logic. In that case, debugging is very time-consuming because no helpful logging is provided.

With this app-initializer logic of angular, we can be sure that the service is initialized before the UI-components are loaded.